kubevirt: move getFlavors() to selectors
